Rename to Elastic.Documentation for now
