+++ Elastic.Documentation/Configuration/Assembler/AssemblyConfiguration.cs
@@ -2,9 +2,10 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
+using Elastic.Documentation.Model.Serialization;
 using YamlDotNet.Serialization;
 
-namespace Documentation.Assembler.Configuration;
+namespace Elastic.Documentation.Model.Configuration.Assembler;
 
 public record AssemblyConfiguration
 {
