+++ Elastic.Documentation/Configuration/Assembler/ContentSource.cs
@@ -6,7 +6,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 using NetEscapades.EnumGenerators;
 
-namespace Elastic.Markdown.IO.State;
+namespace Elastic.Documentation.Model.Configuration.Assembler;
 
 [EnumExtensions]
 public enum ContentSource
