+++ Elastic.Documentation/Configuration/Assembler/PublishEnvironment.cs
@@ -2,10 +2,9 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-using Elastic.Markdown.IO.State;
 using YamlDotNet.Serialization;
 
-namespace Documentation.Assembler.Configuration;
+namespace Elastic.Documentation.Model.Configuration.Assembler;
 
 public record PublishEnvironment
 {
