+++ src/Elastic.Markdown/BuildContext.cs
@@ -5,10 +5,10 @@
 using System.Diagnostics.CodeAnalysis;
 using System.IO.Abstractions;
 using System.Web;
+using Elastic.Documentation.Model;
 using Elastic.Markdown.Diagnostics;
 using Elastic.Markdown.IO;
 using Elastic.Markdown.IO.Configuration;
-using Elastic.Markdown.IO.Discovery;
 using Elastic.Markdown.IO.State;
 
 namespace Elastic.Markdown;
