+++ src/Elastic.Markdown/Slices/HtmlWriter.cs
@@ -4,9 +4,9 @@
 
 using System.Collections.Concurrent;
 using System.IO.Abstractions;
+using Elastic.Documentation.Model;
 using Elastic.Markdown.Extensions.DetectionRules;
 using Elastic.Markdown.IO;
-using Elastic.Markdown.IO.Discovery;
 using Elastic.Markdown.IO.HistoryMapping;
 using Elastic.Markdown.IO.Navigation;
 using Markdig.Syntax;
