+++ src/Elastic.Markdown/SourceGenerationContext.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information
 
 using System.Text.Json.Serialization;
-using Elastic.Markdown.IO.Discovery;
+using Elastic.Documentation.Model;
 using Elastic.Markdown.IO.State;
 using Elastic.Markdown.Links.CrossLinks;
