+++ src/docs-assembler/Building/PublishEnvironmentUriResolver.cs
@@ -3,8 +3,8 @@
 // See the LICENSE file in the project root for more information
 
 using System.Collections.Frozen;
-using Documentation.Assembler.Configuration;
 using Documentation.Assembler.Extensions;
+using Elastic.Documentation.Model.Configuration.Assembler;
 using Elastic.Markdown.Links.CrossLinks;
 
 namespace Documentation.Assembler.Building;
