+++ src/docs-assembler/Cli/InboundLinkCommands.cs
@@ -6,9 +6,9 @@ using System.Diagnostics.CodeAnalysis;
 using System.IO.Abstractions;
 using Actions.Core.Services;
 using ConsoleAppFramework;
+using Elastic.Documentation.Model;
 using Elastic.Documentation.Tooling.Diagnostics.Console;
 using Elastic.Markdown.IO;
-using Elastic.Markdown.IO.Discovery;
 using Elastic.Markdown.Links.InboundLinks;
 using Microsoft.Extensions.Logging;
