+++ src/docs-assembler/Cli/NavigationCommands.cs
@@ -9,10 +9,10 @@ using ConsoleAppFramework;
 using Documentation.Assembler.Building;
 using Documentation.Assembler.Links;
 using Documentation.Assembler.Navigation;
+using Elastic.Documentation.Model;
 using Elastic.Documentation.Tooling.Diagnostics.Console;
 using Elastic.Documentation.Tooling.Filters;
 using Elastic.Markdown.IO;
-using Elastic.Markdown.IO.Discovery;
 using Elastic.Markdown.Links.InboundLinks;
 using Microsoft.Extensions.Logging;
