+++ src/docs-assembler/Cli/RepositoryCommands.cs
@@ -12,10 +12,10 @@ using Amazon.S3;
 using Amazon.S3.Model;
 using ConsoleAppFramework;
 using Documentation.Assembler.Building;
-using Documentation.Assembler.Configuration;
 using Documentation.Assembler.Mapping;
 using Documentation.Assembler.Navigation;
 using Documentation.Assembler.Sourcing;
+using Elastic.Documentation.Model.Configuration.Assembler;
 using Elastic.Documentation.Tooling.Diagnostics.Console;
 using Elastic.Markdown;
 using Elastic.Markdown.Exporters;
