+++ src/docs-builder/Cli/InboundLinkCommands.cs
@@ -6,10 +6,10 @@ using System.Diagnostics.CodeAnalysis;
 using System.IO.Abstractions;
 using Actions.Core.Services;
 using ConsoleAppFramework;
+using Elastic.Documentation.Model;
 using Elastic.Documentation.Tooling.Diagnostics.Console;
 using Elastic.Documentation.Tooling.Filters;
 using Elastic.Markdown.IO;
-using Elastic.Markdown.IO.Discovery;
 using Elastic.Markdown.Links.InboundLinks;
 using Microsoft.Extensions.Logging;
